requirements.txt
numpy==1.18.4
requests==2.18.4
Keras==2.3.1
python-dotenv==0.13.0
